// src/enums/realtime-command.ts
export enum ERealtimeCommand {
  ConnectionStatus = "[Realtime] Connection Status",
  Error = "[Realtime] Error",
  Event = "[Realtime] Event",
  Join = "[Realtime] Join",
  Leave = "[Realtime] Leave",
  Message = "[Realtime] Message",
  PresenceFull = "[Realtime] Presence Full",
  PresenceUpdate = "[Realtime] Presence Update",
}